% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.R
\name{get_md_cr}
\alias{get_md_cr}
\title{Get the metadata for a Crossref snapshot}
\usage{
get_md_cr(
  url = "https://api.crossref.org/snapshots/monthly/latest/all.json.tar.gz"
)
}
\arguments{
\item{url}{The url to the snapshot as documented by Crossref.
Passed on to \code{\link[httr:HEAD]{httr::HEAD()}}.}
}
\description{
Gets metadata from the AWS S3 \href{https://docs.aws.amazon.com/AmazonS3/latest/API/RESTCommonResponseHeaders.html}{response header}.

Happily, this works without authorisation for Metadata Plus.
}
\examples{
get_md_cr()

}
\concept{cr}
\concept{lake}
\concept{version}
