% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leine-package.R
\docType{package}
\name{leine-package}
\alias{leine}
\alias{leine-package}
\title{leine: Data Pipelines for subugoe/wag}
\description{
Provides documentation, glue code and wrappers for the data pipelines
    used at WAG.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://subugoe.github.io/leine}
  \item \url{https://github.com/subugoe/leine}
  \item Report bugs at \url{https://github.com/subugoe/leine/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Other contributors:
\itemize{
  \item Göttingen State and University Library [copyright holder, funder]
}

}
\keyword{internal}
