% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr.R
\name{md_cr_snaps_m}
\alias{md_cr_snaps_m}
\alias{update_md_cr_snaps_m}
\alias{get_md_cr_snaps_m}
\title{Metadata for Crossref monthly Snapshots}
\usage{
md_cr_snaps_m()

update_md_cr_snaps_m()

get_md_cr_snaps_m()
}
\description{
Crossref releases dumps of their database on the 5th of every months.
These dumps sometimes change after the fact.
This metadata, stored in this package,
locks down the versioning of the dumps for our analyses.
}
\details{
To lock down the reproducibility of the metadata dumps,
we store their checksums in this package,
instead of merely relying on their \emph{release date}.
(They have sometimes been updated in the past \emph{after} release.)

Computing a checksum ourselves would be quite expensive
Gets metadata from the AWS S3 \href{https://docs.aws.amazon.com/AmazonS3/latest/API/RESTCommonResponseHeaders.html}{response header}.

Happily, this works without authorisation for Metadata Plus.

Current monthly snapshots:\if{html}{\out{<div class="r">}}\preformatted{knitr::kable(md_cr_snaps_m())
}\if{html}{\out{</div>}}\tabular{llllr}{
   period \tab url \tab date_modified \tab etag \tab size \cr
   2018-05 \tab https://api.crossref.org/snapshots/monthly/2018/05/all.json.tar.gz \tab 2018-06-08 20:32:07 \tab 5fdf4c3e01a713c4391fbc193391a948-5803 \tab 48673539319 \cr
   2018-06 \tab https://api.crossref.org/snapshots/monthly/2018/06/all.json.tar.gz \tab 2018-07-26 20:33:22 \tab 66daedd99bbac1b0406588af39654199-5913 \tab 49595814726 \cr
   2018-07 \tab https://api.crossref.org/snapshots/monthly/2018/07/all.json.tar.gz \tab 2018-08-04 19:15:24 \tab 8b9581b551ac55e255efffed7553dc78-5937 \tab 49800224394 \cr
   2018-08 \tab https://api.crossref.org/snapshots/monthly/2018/08/all.json.tar.gz \tab 2018-09-03 20:04:12 \tab 3250e1960d390c879cacf7ae8fe8e4e0-6006 \tab 50374663277 \cr
   2018-09 \tab https://api.crossref.org/snapshots/monthly/2018/09/all.json.tar.gz \tab 2018-10-11 19:28:42 \tab 48dd8175cf4ef52f8452745629310e79-6153 \tab 51606812435 \cr
   2018-10 \tab https://api.crossref.org/snapshots/monthly/2018/10/all.json.tar.gz \tab 2018-11-03 22:07:04 \tab 2de16d1ccc8fa906aa26f49d2fe752e7-6235 \tab 52296866610 \cr
   2018-11 \tab https://api.crossref.org/snapshots/monthly/2018/11/all.json.tar.gz \tab 2018-12-03 22:02:12 \tab 123a2132ede3991a4bee12d96e4ae51c-6169 \tab 51747932929 \cr
   2018-12 \tab https://api.crossref.org/snapshots/monthly/2018/12/all.json.tar.gz \tab 2019-01-05 01:24:48 \tab 250e3e72cc5bd4ab5dbec899b8db7580-6267 \tab 52564041493 \cr
   2019-01 \tab https://api.crossref.org/snapshots/monthly/2019/01/all.json.tar.gz \tab 2020-02-19 19:19:18 \tab 6305e97bb00c764b1102d22c2be6d663-8495 \tab 71258741599 \cr
   2019-02 \tab https://api.crossref.org/snapshots/monthly/2019/02/all.json.tar.gz \tab 2019-03-04 14:52:26 \tab dd03bf567aed0e41b3736d9e9f607e85-6647 \tab 55758917898 \cr
   2019-03 \tab https://api.crossref.org/snapshots/monthly/2019/03/all.json.tar.gz \tab 2019-04-03 14:16:02 \tab fe2cd8dd764503640f74806e34248781-6840 \tab 57377765958 \cr
   2019-04 \tab https://api.crossref.org/snapshots/monthly/2019/04/all.json.tar.gz \tab 2019-05-03 23:16:04 \tab 160dc0f8d12b3c91c6ed2ad0df87b4e4-7170 \tab 60145344184 \cr
   2019-05 \tab https://api.crossref.org/snapshots/monthly/2019/05/all.json.tar.gz \tab 2019-06-04 12:16:05 \tab 1eb34559b0e80b6ad778c7a40d1dd45e-7623 \tab 63942201992 \cr
   2019-06 \tab https://api.crossref.org/snapshots/monthly/2019/06/all.json.tar.gz \tab 2019-07-13 16:05:50 \tab 586e1ace0577ca44c422a0029960098b-7773 \tab 65197539697 \cr
   2019-07 \tab https://api.crossref.org/snapshots/monthly/2019/07/all.json.tar.gz \tab 2019-08-04 05:46:04 \tab 31af31381ff92756beb7ab7559c33291-7858 \tab 65912914209 \cr
   2019-08 \tab https://api.crossref.org/snapshots/monthly/2019/08/all.json.tar.gz \tab 2019-09-04 07:46:04 \tab 17782fc0071ee11ff5db519446e1edf0-7945 \tab 66646812608 \cr
   2019-09 \tab https://api.crossref.org/snapshots/monthly/2019/09/all.json.tar.gz \tab 2019-10-04 12:42:40 \tab 196ea8897b159c0378dc5086d6fca861-8043 \tab 67464925959 \cr
   2019-10 \tab https://api.crossref.org/snapshots/monthly/2019/10/all.json.tar.gz \tab 2019-11-13 13:42:20 \tab 5eef0f01562f7bc31d5a016fbd6042f7-5933 \tab 99530559599 \cr
   2019-11 \tab https://api.crossref.org/snapshots/monthly/2019/11/all.json.tar.gz \tab 2019-12-05 17:16:04 \tab 3e417701f575d2a9a19582d2606231d2-8307 \tab 69683786041 \cr
   2019-12 \tab https://api.crossref.org/snapshots/monthly/2019/12/all.json.tar.gz \tab 2020-01-06 20:39:19 \tab 66db92cdd582f8cf93b03e9b3ade167d-8424 \tab 70662719003 \cr
   2020-01 \tab https://api.crossref.org/snapshots/monthly/2020/01/all.json.tar.gz \tab 2020-02-19 19:40:08 \tab 6305e97bb00c764b1102d22c2be6d663-8495 \tab 71258741599 \cr
   2020-02 \tab https://api.crossref.org/snapshots/monthly/2020/02/all.json.tar.gz \tab 2020-03-05 10:21:09 \tab cf5dbcfb9f8c4e01d7dd0d718429b642-8558 \tab 71782804006 \cr
   2020-03 \tab https://api.crossref.org/snapshots/monthly/2020/03/all.json.tar.gz \tab 2020-04-04 08:31:09 \tab 47b39e33efd9824c7c5c68f06021ba88-8658 \tab 72624782825 \cr
   2020-04 \tab https://api.crossref.org/snapshots/monthly/2020/04/all.json.tar.gz \tab 2020-05-11 11:52:41 \tab 104df834ffb8ff0dfd20e121a6318618-8823 \tab 74010463525 \cr
   2020-05 \tab https://api.crossref.org/snapshots/monthly/2020/05/all.json.tar.gz \tab 2020-06-04 13:31:09 \tab aeafd8a98336812226affb0bedc973cd-8908 \tab 74724383102 \cr
   2020-06 \tab https://api.crossref.org/snapshots/monthly/2020/06/all.json.tar.gz \tab 2020-07-04 12:31:08 \tab f5fbc47585eb76f74887decf4b4d36e7-9013 \tab 75604365177 \cr
   2020-07 \tab https://api.crossref.org/snapshots/monthly/2020/07/all.json.tar.gz \tab 2020-08-04 16:31:04 \tab 1e56d469216fc77d07d103a15d9d7427-9120 \tab 76498078549 \cr
   2020-08 \tab https://api.crossref.org/snapshots/monthly/2020/08/all.json.tar.gz \tab 2020-09-05 17:34:33 \tab 5cb71554733ac30fcb95c0b99d4c7138-9230 \tab 77423432433 \cr
   2020-09 \tab https://api.crossref.org/snapshots/monthly/2020/09/all.json.tar.gz \tab 2020-10-27 21:16:29 \tab 1936b0825dec4b11db705f8053b7da78-9415 \tab 78973357479 \cr
   2020-10 \tab https://api.crossref.org/snapshots/monthly/2020/10/all.json.tar.gz \tab 2020-11-05 06:01:07 \tab ab52c1f7a2235692faa34d431751badb-9451 \tab 79275265379 \cr
   2020-11 \tab https://api.crossref.org/snapshots/monthly/2020/11/all.json.tar.gz \tab 2020-12-05 09:31:09 \tab 7216bfd6eafcd0ff83894acece592c29-9556 \tab 80153941136 \cr
   2020-12 \tab https://api.crossref.org/snapshots/monthly/2020/12/all.json.tar.gz \tab 2021-01-12 16:46:08 \tab 8f0977e27c695a954f843503a6120946-6367 \tab 106818152559 \cr
   2021-01 \tab https://api.crossref.org/snapshots/monthly/2021/01/all.json.tar.gz \tab 2021-02-05 17:01:08 \tab 242406bc4cce046eaae133f4cf648978-6482 \tab 108733400373 \cr
   2021-02 \tab https://api.crossref.org/snapshots/monthly/2021/02/all.json.tar.gz \tab 2021-03-05 18:31:09 \tab e259608e23b932b6aed7f7c67b5256b9-6564 \tab 110120778006 \cr
   2021-03 \tab https://api.crossref.org/snapshots/monthly/2021/03/all.json.tar.gz \tab 2021-04-05 17:31:08 \tab 9a730fb1461b0f170940845fbb93dd2f-6642 \tab 111419433519 \cr
}
}
\section{Related Functions and Methods}{
\subsection{Functions}{
\itemize{
\item \code{update_md_cr_snaps_m}: Update the metadata of monthly snapshots stored in leine
}
\itemize{
\item \code{get_md_cr_snaps_m}: Get metadata for currently available monthly snapshots
}}}

\concept{cr}
\concept{lake}
\concept{version}
